refactor(automation): decompose CIDriver god-class into 4 SRP collaborators (#1289)
Extract PRDiscovery, CICheckInspector, CIFixOrchestrator, and PostMergeProcessor
from ci_driver.py (3 358 → 1 410 lines). Each collaborator owns one narrow
concern; CIDriver retains thin delegation stubs so all patch.object(CIDriver, …)
test targets remain valid. Shared mutable state is threaded via provider
callables and a __setattr__ hook propagates state_dir changes to collaborators.
Structural tests updated to scan companion modules alongside ci_driver.py.
