+++ tests/unit/automation/test_invoke_allowed_tools_scoping.py
@@ -18,7 +18,9 @@ CALL_SITES = [
     ("plan_reviewer.py", {"Read", "Glob", "Grep"}, False),
     ("review_validator.py", {"Read", "Glob", "Grep"}, False),
     ("planner_claude.py", {"Read", "Glob", "Grep", "Bash"}, True),
-    ("ci_driver.py", {"Read", "Glob", "Grep", "Bash"}, True),
+    # #1289: invoke_claude_with_session calls moved to SRP collaborator modules.
+    ("ci_fix_orchestrator.py", {"Read", "Glob", "Grep", "Bash"}, True),
+    ("post_merge_processor.py", {"Read", "Write", "Edit", "Glob", "Grep", "Bash"}, True),
     ("implementer_phase_runner.py", {"Read", "Glob", "Grep", "Bash"}, True),
     ("comment_difficulty.py", {"Read", "Glob", "Grep"}, False),
 ]
